update name of data save file and combine in data getter
